requirements.txt
Flask==3.0.0
pandas==2.1.4
xhtml2pdf==0.2.16
openpyxl==3.1.2
Jinja2==3.1.2
reportlab==4.0.9
arabic-reshaper==3.0.0
python-bidi==0.4.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
